css changes in group
