+++ src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import { Login, Game, Feedback, Ranking, Settings } from './pages';
+import './App.css';
 
 export default function App() {
   return (
