+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as GameBoard } from './GameBoard';
+export { default as Header } from './Header';
