+++ src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 import store from './redux/store';
-import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
