+++ src/pages/Game/Game.module.css
@@ -1,0 +1,3 @@
+.main {
+  background: var(--primary);
+}
